All the required work : Consommation
This is a new full sql modified database, this is the last one don't delete it because the structure is modified .
The main page is index.php accompanied with style.css. The option list extracts from Database
consommation.php is the main php file for adding in the database. consommation2.php is the a second version written in php mysqli (not necessary to include in the project but for tests reasons).